Add ProcessRegistrySubscriber (#10463)
* Inline RuleChangedListener

* Rename ScalingRegistryService

* Rename ScalingRegistryService

* Add ProcessRegistrySubscriber
